Use constructor parameters as source of definitions
Constructor promoted properties provide better flexibility. However,
it's not possible to get their default value as it is part of the
constructor parameter (which assigns the property).

I decided that all inputs need to define a constructor with list of
parameters. Those don't need to be promoted properties, but all
parameters with #[Argument], or #[Option] will be used to create the
definitions.